+++ GuZhenRenCode/Cards/Core/AbstractShaZhaoCard.cs
@@ -1,4 +1,6 @@
-﻿using MegaCrit.Sts2.Core.Commands;
+﻿using GuZhenRen.Multiplayer;
+
+using MegaCrit.Sts2.Core.Commands;
 using MegaCrit.Sts2.Core.Entities.Cards;
 using MegaCrit.Sts2.Core.Entities.Players;
 using MegaCrit.Sts2.Core.GameActions.Multiplayer;
@@ -217,6 +219,9 @@ public abstract class AbstractShaZhaoCard
                 item => item.card.Enchantment?.Id.ToString() ??
                     string.Empty,
                 StringComparer.Ordinal
+            )
+            .ThenBy(item =>
+                GuZhenRenDeterminism.GetCardNetworkId(item.card)
             )
             .ThenBy(item => item.index)
             .Select(item => item.card)